+++ sit-rdg-core/src/main/java/io/sitoolkit/rdg/core/domain/generator/TableDataGenerator.java
@@ -34,7 +34,7 @@ public abstract class TableDataGenerator {
     return table.getFullyQualifiedName();
   }
 
-  public int getRequiredRowCount() {
+  public long getRequiredRowCount() {
     return config.getRowCount(table);
   }
